trying to fix this test (#70)

+++ Quiz-app/fxui/src/test/java/quizapp/ui/AddQuizControllerTest.java
@@ -1,26 +1,21 @@
 package quizapp.ui;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestMethodOrder;
+import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
+import java.awt.*;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.Label;
 import javafx.stage.Stage;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
-import quizapp.core.Quiz;
 import javafx.scene.control.MenuButton;
 import javafx.scene.control.ScrollPane;
-
-import java.awt.*;
 import javafx.scene.control.TextField;
-
-
-import org.junit.jupiter.api.Order;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestMethodOrder;
-import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
+import quizapp.core.Quiz;
 
 @TestMethodOrder(OrderAnnotation.class)
 public class AddQuizControllerTest extends FxuiTest {
@@ -38,14 +33,20 @@ public class AddQuizControllerTest extends FxuiTest {
     stage.setScene(new Scene(root));
     stage.show();
     this.stage = stage;
+    setUp();
   }
 
+  public void setUp() throws AWTException {
+    fillInQuiz();
+  }
 
   @Test
-  @Order(1)
+  //@Order(1)
   public void MakeAQuizTest() throws AWTException {
-    //Deletes quiz if it already exists
+    //Deletes quiz if it already exist
+    if (directQuizAccess.getQuiz("Color-quiz") != null) {
     directQuizAccess.deleteQuiz("Color-quiz");
+    }
     fillInQuiz();
     Quiz quiz = directQuizAccess.getQuiz("Color-quiz");
     assertNotNull(quiz);
@@ -56,10 +57,11 @@ public class AddQuizControllerTest extends FxuiTest {
     // Checks that the scene has changed to main menu
     assertNull(stage.getScene().lookup("#mainMenu"));
     assertNotNull(stage.getScene().lookup("#menuButton"));
+    directQuizAccess.deleteQuiz("Color-quiz");
   }
 
   @Test
-  @Order(2)
+  //@Order(2)
   public void submitUncompleteQuizTest() throws AWTException {
     Robot r = new Robot();
     findTextField("#title").setText("Test quiz");
@@ -84,17 +86,18 @@ public class AddQuizControllerTest extends FxuiTest {
   }
 
   @Test
-  @Order(3)
+  //@Order(3)
   public void makeQuizWithInvalidName() throws AWTException {
-    fillInQuiz();
+    fillInQuizWithSameName();
     String text = ((Label) stage.getScene().lookup("#scoreLabel")).getText();
     assertEquals("Invalid Quizname. The title must be unique, there is already a quiz named Color quiz", text);
+    directQuizAccess.deleteQuiz("Color-quiz");
   }
 
 
 
   @Test
-  @Order(4)    
+  //@Order(4)    
   public void goToMainMenuTest() {
     // Checks that we on add quiz page
     assertNotNull(stage.getScene().lookup("#mainMenu"));
@@ -108,7 +111,7 @@ public class AddQuizControllerTest extends FxuiTest {
 
 
   @Test
-  @Order(5)    
+  //@Order(5)    
   public void checkCorrectUserDisplayed() {
     // Checks active user and makes sure it matches username displayed on menu button
     String activeUser = ((MenuButton)stage.getScene().lookup("#userMenu")).getText();
@@ -158,6 +161,52 @@ public class AddQuizControllerTest extends FxuiTest {
     findTextField("#q2an2").setText("Purple");
     findTextField("#q2an3").setText("Green");
     clickOnButton("#q2a2");
+    try {
+      Thread.sleep(100);
+    } catch (InterruptedException e) {
+    }
+    clickOnButton("#submit");
+  }
+
+  private void fillInQuizWithSameName() throws AWTException {
+    ScrollPane scroll = (ScrollPane) stage.getScene().lookup("#scroll");
+    // Fills in everything needed to make a quiz
+    scroll.setVvalue(0.01);
+    findTextField("#title").setText("Color quiz");
+    findTextField("#q0").setText("What color do you get if you mix red and yellow?");
+    findTextField("#q0an0").setText("Blue");
+    findTextField("#q0an1").setText("Orange");
+    findTextField("#q0an2").setText("Purple");
+    findTextField("#q0an3").setText("Red");
+    clickOnButton("#q0a1");
+    try {
+      Thread.sleep(100);
+    } catch (InterruptedException e) {
+    }
+    scroll.setVvalue(0.5);
+    try {
+      Thread.sleep(1000);
+    } catch (InterruptedException e) {
+    }
+    findTextField("#q1").setText("What color do you get if you mix red and blue?");
+    findTextField("#q1an0").setText("Green");
+    findTextField("#q1an1").setText("Yellow");
+    findTextField("#q1an2").setText("Blue");
+    findTextField("#q1an3").setText("Purple");
+    clickOnButton("#q1a3");
+    // Scrolls to bottom of screen
+    // Slows down the code to give the robot time to scroll
+    scroll.setVvalue(1);
+     try {
+      Thread.sleep(1000);
+    } catch (InterruptedException e) {
+    }
+    findTextField("#q2").setText("What color do you get if you mix green and green?");
+    findTextField("#q2an0").setText("Black");
+    findTextField("#q2an1").setText("White");
+    findTextField("#q2an2").setText("Purple");
+    findTextField("#q2an3").setText("Green");
+    clickOnButton("#q2a3");
     try {
       Thread.sleep(100);
     } catch (InterruptedException e) {
